Fix test failure due to use of half type

+++ tests/"quotes".cl
@@ -4,7 +4,7 @@ __constant int test_array1[__LINE__ == 3 ? 1 : -1] = {0};
 __constant int test_array2[__FILE__[9] == '"' ? 1 : -1] = {0};
 __constant int test_array3[__FILE__[16] == '"' ? 1 : -1] = {0};
 
-__kernel void main(__global const half *in, __global half *out)
+__kernel void main(__global const int *in, __global int *out)
 {
     *out = *in;
 }
